requirements.txt
moviepy==1.0.3
praw==7.1.0
gTTS==2.2.1
mutagen==1.45.1
Unihandecode==0.81
Pillow==8.0.1
